pyproject.toml
[tool.poetry]
name = "MeowerBot"
version = "3.2.0"
description = "A meower bot lib for py"
authors = ["showierdata9978 <68120127+showierdata9978@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "MeowerBot"}
]
repository = "https://github.com/MeowerBots/MeowerBot.py"


[tool.poetry.dependencies]
python = "^3.12"
websockets = "^11.0.3"
orjson = "^3.9.4"
httpx = "^0.25.0"
dataclasses-json = "^0.6.1"
ujson = "^5.8.0"



[tool.poetry.group.dev.dependencies]

types-ujson = "^5.8.0.1"
flake8 = "^6.1.0"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core>=1.0.0", ]
build-backend = "poetry.core.masonry.api"
